Fix iterating for loop with vector size_t

+++ util.h
@@ -7,7 +7,7 @@
 template<typename T>
 void printVector(std::vector<T>& vec) {
   std::cout << "[";
-  for (int i = 0; i < vec.size(); i++) {
+  for (size_t i = 0; i < vec.size(); i++) {
     if (i != vec.size() - 1) std::cout << vec[i] << ", ";
     else std::cout << vec[i];
   }
@@ -18,7 +18,7 @@ template<typename T>
 void print2DVector(std::vector<std::vector<T>>& vec) {
   for (auto v : vec) {
     std::cout << "[";
-    for (int i = 0; i < v.size(); i++) {
+    for (size_t i = 0; i < v.size(); i++) {
       if (i != v.size() - 1) std::cout << v[i] << ", ";
       else std::cout << v[i];
     }
